Returned solutionid to a 32-char string
