+++ packages/common/src/generalHelpers.ts
@@ -20,6 +20,7 @@
  * @module generalHelpers
  */
 
+import { createId } from "@esri/hub-common";
 import { IDatasourceInfo, IItemTemplate } from "./interfaces";
 import { Sanitizer, sanitizeJSON } from "./libConnectors";
 import { new_File } from "./polyfills";
@@ -103,6 +104,34 @@ export function blobToText(blob: Blob): Promise<string> {
  */
 export function checkUrlPathTermination(url: string): string {
   return url ? (url.endsWith("/") ? url : url + "/") : url;
+}
+
+/**
+ * Creates a random 32-character alphanumeric string.
+ * 
+ * @return A lowercase 32-char alphanumeric string
+ * @internal
+ */
+export function createLongId(): string {
+  // createId gets a random number, converts it to base 36 representation, then grabs chars 2-8
+  return createId() + createId() + createId() + createId();
+}
+
+/**
+ * Creates a random 8-character alphanumeric string that begins with an alphabetic character.
+ *
+ * @return An alphanumeric string in the range [a0000000..zzzzzzzz]
+ */
+export function createShortId(): string {
+  // Return a random number, but beginning with an alphabetic character so that it can be used as a valid
+  // dotable property name. Used for unique identifiers that do not require the rigor of a full UUID -
+  // i.e. node ids, process ids, etc.
+  const min = 0.2777777777777778; // 0.a in base 36
+  const max = 0.9999999999996456; // 0.zzzzzzzz in base 36
+  return (_getRandomNumberInRange(min, max).toString(36) + "0000000").substr(
+    2,
+    8
+  );
 }
 
 /**
@@ -650,6 +679,20 @@ export function getTemplateById(templates: IItemTemplate[], id: string): any {
 }
 
 // ------------------------------------------------------------------------------------------------------------------ //
+
+/**
+ * Creates a random number between two values.
+ *
+ * @param min Inclusive minimum desired value
+ * @param max Non-inclusive maximum desired value
+ * @return Random number in the range [min, max)
+ */
+export function _getRandomNumberInRange(min: number, max: number): number {
+  // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/random#Getting_a_random_number_between_two_values
+  // © 2006 IvanWills
+  // MIT license https://opensource.org/licenses/mit-license.php
+  return Math.random() * (max - min) + min;
+}
 
 /**
  * Pads the string representation of a number to a minimum width. Requires modern browser.
